hw6/requirements.txt
Flask==1.0.2
Cerberus
json2html
requests
google-cloud-datastore==1.7.3
